Add files via upload
Data Description:
This dataset contains information extracted from PDF files available on the website https://www.chospab.es/biblioteca/.

Disclaimer:
The data provided in this dataset is derived from PDF files publicly available on the website mentioned above. We do not guarantee the accuracy, completeness, or reliability of the information obtained through this extraction process. Additionally, we are not responsible for any misuse or consequences arising from the use of this data. Users are advised to verify the information from the original PDF files or the source website before making any decisions based on it.